flask-task-api/requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
PyJWT==2.8.0
Werkzeug==3.0.3
pytest==8.1.1
